Remove on chain storage of geoPoints
Storing data on chain is very costly, but an important part of creating immutable ledgers.

With EcoBux, we want the most money we raise to go to the charities we partner with. This is why we are changing the onchain storage to enable more money going to the charities.

This saves EcoBux more than $1,000 at the time of this commit, money better spent in the hands of charities.

This change can be easily reverted if need be.